Removed testbench mappings and instead added testbench control via the dutcontrol function
The dutcontrol function applies the correct output portmaps due to the function being overloaded for every gate created so far in the project.

ALSO:

Added 4:2 Encoder gate to the project which was the reason dutcontrol was needed since it returns a bit vector instead of just a single bit output.